Added JWT tokens - API protection

+++ prototipo1/backend/app.js
@@ -8,28 +8,30 @@ const crypto = require('crypto');
 const path = require('path');
 const mongo = require('mongodb');
 const bcrypt=require("bcrypt")
-const session=require("express-session")
 const MongoStore= require("connect-mongo")
 const app = express(); //intacia de express
 const https = require('https');
+const jwt=require('jsonwebtoken');
+const cookieParser=require('cookie-parser');
 
 app.use(bodyParser.json()); // definir body-parser como json
 app.use(bodyParser.urlencoded({ extended: true })); // decirle a la app que use body-parser
-app.use(cors()); // decirle a la app que use cors
+app.use(cookieParser())
+
+// app.use(cors()); // decirle a la app que use cors
+var corsOptions = {
+    origin: '*',
+    credentials: true,
+    exposedHeaders: ["set-cookie"],
+};
+    
+app.use(cors(corsOptions));
 
 const DB_URL = "mongodb://127.0.0.1:27017/demoDB";
-
-app.use(session({
-    secret: "secret",
-    resave: false,
-    saveUninitialized: false,
-    store: MongoStore.create({mongoUrl: DB_URL})
-}))
 
 const uploads = multer({dest: ".temp"}) // definir multer para que guarde los archivos en la carpeta .temp
 
 function connectToDB(){
-    //let client = new MongoClient("mongodb://localhost/demoDB2"); // string de conexion dice que vamos a usar mongo en localhost (default) y la db demoDB2
     let client = new MongoClient(DB_URL);
     client.connect();
     db = client.db();
@@ -43,40 +45,46 @@ app.post("/login", (req, res)=>{
       if(result!=null){
         bcrypt.compare(pass, result.password, (err, resultB)=>{
           if(resultB){
-            res.json(result)
-            // req.session.usuario=user;
-            // res.redirect('/home');
+            let token = jwt.sign({
+                usuario:user,
+                exp: Math.floor(Date.now() / 1000) + (60 * 10),
+                iat: Math.floor(Date.now())
+              },'secretKey')
+            res.json(token)
+            // res.cookie("token", token)
+            // res.json(result) // si las credenciales son correctas, regresar info del usuario
           }else{
             console.log("Credenciales Incorrectas")
-            res.json(null)
-            // res.redirect('http://localhost:3000/login');
+            res.json(null) //indicar en front credenciales incorrectas
           }
         })
       }else{
         console.log("Credenciales Incorrectas")
-        res.json(null)
-        // res.redirect('http://localhost:3000/login');
+        res.json(null) //indicar en front credenciales incorrectas
       }
     })
   })
 
-  
-// app.get("/logout", (req, res)=>{
-//     req.session.destroy();
-//     res.redirect("/home");
-// })
-
-
-app.get("/", function(req, res){
-    db.collection("nulidad").find({}).toArray(function(err, result){
+  app.get("/nulidad", function(req, res){
+    jwt.verify(req.headers.token, "secretKey", (err, userId) => {
         if(err){
-            handleError(res, err.message, "Failed to get nulidad");
-        }else{
-            res.status(200).json(result);
-            
+            res.json(null) 
         }
-    });
+        else {
+            console.log(userId)
+            db.collection("nulidad").find({}).toArray(function(err, result){
+                if(err){
+                    handleError(res, err.message, "Failed to get nulidad");
+                }else{
+                    res.status(200).json(result);
+                    
+                }
+            });
+
+        }
+    })    
 })
+
 
 app.post("/crearExpediente", function(req, res){
     let aInsertar = {nombre:req.body.nombre, numero:req.body.numero, expediente:req.body.expediente, actor:req.body.actor, estatus:req.body.estatus, fecha:req.body.fecha}
@@ -91,44 +99,48 @@ app.post("/crearExpediente", function(req, res){
 
 //Mostrar archivos de un expediente en especifico
 app.get("/descargarArchivos", function(req, res){
-    db.collection("pruebaUpload").find({ "expediente": req.query.expediente }).toArray(function(err, result){
-        if(err){
-            handleError(res, err.message, "Failed to get documentos de expediente");
-        }else{
-            res.status(200).json(result);
-        }
-    }); 
+        db.collection("pruebaUpload").find({ "expediente": req.query.expediente }).toArray(function(err, result){
+            if(err){
+                handleError(res, err.message, "Failed to get documentos de expediente");
+            }else{
+                res.status(200).json(result);
+            }
+        }); 
 });
 
 //Descargar archivos por el nombre
 app.get("/descargarArchivos/download", async (req, res) => {
-    console.log(req.query);
-    db.collection("pruebaUpload").findOne({"_id": mongo.ObjectId(req.query.id)}, function(err, result){
-        if(err){
-            handleError(res, err.message, "Failed to get documentos de expediente");
-        }else{
-            console.log(__dirname + "/.storage/" + req.query.nombre);
-            let temporal = __dirname + "/.temp/" + req.query.nombre + ".pdf";
-            let inputFS = fs.createReadStream(__dirname + result.archivo)
-            let outputFS = fs.createWriteStream(temporal)
-
-            let key = "abcabcabcabcabcabcabcabcabcabc12" //Not to be done, key usually goes encrypted as well 32 bytes(256 bits) 
-            let iv = "abcabcabcabcabc1" //(initialization vector) 16 bytes
-
-            let cipher = crypto.createDecipheriv("aes-256-cbc", key, iv)
-            inputFS.pipe(cipher).pipe(outputFS)
-            outputFS.on("finish", ()=>{
-                console.log(temporal);
-                res.set({
-                    'Content-Type': 'application/pdf'
+    if(req.headers.referer){
+        db.collection("pruebaUpload").findOne({"_id": mongo.ObjectId(req.query.id)}, function(err, result){
+            if(err){
+                handleError(res, err.message, "Failed to get documentos de expediente");
+            }else{
+                console.log(__dirname + "/.storage/" + req.query.nombre);
+                let temporal = __dirname + "/.temp/" + req.query.nombre + ".pdf";
+                let inputFS = fs.createReadStream(__dirname + result.archivo)
+                let outputFS = fs.createWriteStream(temporal)
+    
+                let key = "abcabcabcabcabcabcabcabcabcabc12" //Not to be done, key usually goes encrypted as well 32 bytes(256 bits) 
+                let iv = "abcabcabcabcabc1" //(initialization vector) 16 bytes
+    
+                let cipher = crypto.createDecipheriv("aes-256-cbc", key, iv)
+                inputFS.pipe(cipher).pipe(outputFS)
+                outputFS.on("finish", ()=>{
+                    console.log(temporal);
+                    res.set({
+                        'Content-Type': 'application/pdf'
+                    })
+                    res.download(temporal, function(err){
+                        if(err) throw err;
+                        fs.unlinkSync(temporal);
+                    })
                 })
-                res.download(temporal, function(err){
-                    if(err) throw err;
-                    fs.unlinkSync(temporal);
-                })
-            })
-        }
-    }); 
+            }
+        }); 
+    }
+    else{
+        res.send("Unauthorized Access");
+    }
 });
 
 app.post("/subirArchivo", uploads.single("archivo"), (req, res)=>{
@@ -166,9 +178,14 @@ app.get("/tablaCuentas", function(req, res) {
 })
 
 app.delete("/borrarCuenta", function(req, res) {
-    db.collection("accounts").deleteOne({"_id": mongo.ObjectId(req.query.id)}, (err, result) => {
-        if (err) throw err;
-    })
+    if(req.headers.referer){
+        db.collection("accounts").deleteOne({"_id": mongo.ObjectId(req.query.id)}, (err, result) => {
+            if (err) throw err;
+        })
+    }
+    else{
+        res.send("Unauthorized Access");
+    }
 })
  
 app.post("/crearCuenta", (req, res)=>{
